+++ cucumber-expressions/javascript/test/CucumberExpressionTest.ts
@@ -285,19 +285,18 @@ describe('CucumberExpression', () => {
   })
 
   describe('escapes special characters', () => {
-    ;['\\', '[', ']', '^', '$', '.', '|', '?', '*', '+'].forEach(
-      (character) => {
-        it(`escapes ${character}`, () => {
-          const expr = `I have {int} cuke(s) and ${character}`
-          const expression = new CucumberExpression(
-            expr,
-            new ParameterTypeRegistry()
-          )
-          const arg1 = expression.match(`I have 800 cukes and ${character}`)[0]
-          assert.strictEqual(arg1.getValue(null), 800)
-        })
-      }
-    )
+    const special = ['\\', '[', ']', '^', '$', '.', '|', '?', '*', '+']
+    special.forEach((character) => {
+      it(`escapes ${character}`, () => {
+        const expr = `I have {int} cuke(s) and ${character}`
+        const expression = new CucumberExpression(
+          expr,
+          new ParameterTypeRegistry()
+        )
+        const arg1 = expression.match(`I have 800 cukes and ${character}`)[0]
+        assert.strictEqual(arg1.getValue(null), 800)
+      })
+    })
 
     it(`escapes .`, () => {
       const expr = `I have {int} cuke(s) and .`