+++ react/javascript/src/components/gherkin/Background.tsx
@@ -1,21 +1,26 @@
 import React from 'react'
-import Keyword from './Keyword'
 import Description from './Description'
 import { messages } from '@cucumber/messages'
 import StepList from './StepList'
+import IdGenerator from '../../IdGenerator'
+import BackgroundTitle from './BackgroundTitle'
 import IBackground = messages.GherkinDocument.Feature.IBackground
 
 interface IProps {
   background: IBackground
 }
 
+const generator = new IdGenerator()
+
 const Background: React.FunctionComponent<IProps> = ({ background }) => {
+  const idGenerated = generator.generate(background.name)
+
   return (
     <section>
-      <h2>
-        <Keyword>{background.keyword}:</Keyword>{' '}
-        <span className="step-text">{background.name}</span>
-      </h2>
+      <BackgroundTitle
+          id={idGenerated}
+          background={background}
+      />
       <div className="indent">
         <Description description={background.description} />
         <StepList