+++ react/javascript/src/components/gherkin/DocString.tsx
@@ -1,13 +1,18 @@
 import React from 'react'
 import { messages } from '@cucumber/messages'
 import IDocString = messages.GherkinDocument.Feature.Step.IDocString
+import HighLight from '../app/HighLight'
 
 interface IProps {
   docString: IDocString
 }
 
 const DocString: React.FunctionComponent<IProps> = ({ docString }) => {
-  return <pre>{docString.content}</pre>
+  return (
+    <pre>
+      <HighLight text={docString.content} />
+    </pre>
+  )
 }
 
 export default DocString