+++ react/javascript/src/components/gherkin/Tags.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { messages } from '@cucumber/messages'
 import ITag = messages.GherkinDocument.Feature.ITag
+import HighLight from '../app/HighLight'
 
 interface IProps {
   tags: ITag[]
@@ -14,7 +15,7 @@ const Tags: React.FunctionComponent<IProps> = ({ tags }) => {
     <ul className="tag-list">
       {tags.map((tag, index) => (
         <li className="tag" key={index}>
-          {tag.name}
+          <HighLight text={tag.name} />
         </li>
       ))}
     </ul>